+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+
+- **added**: `clear_{param,query_param,fragment}` methods for `PathBuilder`.
+
 ## 0.1.0
 
 - Initial release.